added edit button to HabitDetails widget
