+++ lib/pages/habit_detail_page.dart
@@ -5,6 +5,7 @@ import 'package:habit_tracker/models/habit.dart';
 import 'package:habit_tracker/widgets/custom_card.dart';
 import 'package:habit_tracker/widgets/custom_scaffold.dart';
 import 'package:habit_tracker/widgets/daily_habit.dart';
+import 'package:habit_tracker/widgets/habit_detail.dart';
 import 'package:habit_tracker/widgets/spacer_box.dart';
 import 'package:habit_tracker/widgets/streak_card.dart';
 
@@ -22,59 +23,9 @@ class HabitDetailPage extends StatelessWidget {
       builder: (context, state) {
         Habit habit = state.habits[habitId];
         return CustomScaffold(
-          child: habit != null
-              ? Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      habit.name,
-                      style: Theme.of(context).textTheme.headline3,
-                    ),
-                    SpacerBox.size16(),
-                    _multiDayOverview(habit),
-                    SpacerBox.size16(),
-                    StreakCard(
-                      streak: habit.streak,
-                    ),
-                    SpacerBox.size16(),
-                    _deleteButton(context, habit),
-                  ],
-                )
-              : Container(),
+          child: habit != null ? HabitDetail(habit: habit) : Container(),
         );
       },
     );
-  }
-
-  Widget _deleteButton(BuildContext context, Habit habit) {
-    return Center(
-      child: FlatButton(
-          onPressed: () {
-            Navigator.pop(context);
-            BlocProvider.of<HabitBloc>(context).add(HabitDeleted(habit));
-          },
-          child: Text('delete this habit'.toUpperCase())),
-    );
-  }
-
-  Widget _multiDayOverview(Habit habit) {
-    return CustomCard(
-        padding: 24,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: dates()
-              .map((DateTime date) => DailyHabit(
-                    habit: habit,
-                    date: date,
-                  ))
-              .toList(),
-        ));
-  }
-
-  List<DateTime> dates() {
-    DateTime today = DateTime.now();
-    List<DateTime> dates = List<DateTime>.generate(
-        5, (index) => today.subtract(Duration(days: index)));
-    return dates.reversed.toList();
   }
 }
